+++ Accounting/Models/UserRequest.cs
@@ -7,7 +7,7 @@ namespace Accounting.Models
         [Display(Name = "کد ملی")]
         [Required(ErrorMessage = "لطفا {0} را وارد کنید")]
         [RegularExpression(@"[0-9]{10}", ErrorMessage = "مقدار {0} نامعتبر می باشد")]
-        public decimal NationalCode { get; set; }
+        public long NationalCode { get; set; }
 
         [Display(Name = "رایانامه")]
         [EmailAddress(ErrorMessage = "لطفا {0} معتبری را وارد کنید")]
@@ -16,6 +16,6 @@ namespace Accounting.Models
         [Display(Name = "موبایل")]
         [Required(ErrorMessage = "لطفا {0} را وارد کنید")]
         [RegularExpression(@"[0-9]{10}", ErrorMessage = "مقدار {0} نامعتبر می باشد، می بایست بدون صفر و پیش کد وارد گردد")]
-        public decimal? Mobile { get; set; }
+        public long? Mobile { get; set; }
     }
 }
